backend/requirements.txt

fastapi
uvicorn[standard]
python-multipart
pydantic
pydantic-settings


firebase-admin

python-jose[cryptography]
passlib[bcrypt]

pypdf2
python-docx
python-magic

pinecone-client

openai
anthropic
google-generativeai

langchain
langchain-openai
langchain-anthropic
tiktoken
nltk


python-dotenv
httpx
pytest
pytest-asyncio
